Feat: 채팅 input 하단에 안내 메시지 추가 (#48)
