+++ src/features/chatting/containers/ChattingContainer.tsx
@@ -1,19 +1,17 @@
 import { type ReactNode, type Dispatch, type SetStateAction, Fragment, useRef } from "react";
-import { Platform, ScrollView } from "react-native";
-import { Image as Img } from "expo-image";
+import { ScrollView } from "react-native";
+
 import styled from "styled-components/native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import Loading from "@/src/components/ui/Loading";
 import CommonError from "@/src/components/ui/CommonError";
-import CachedImage from "@/src/components/ui/CachedImage";
-import { shadowProps } from "@/src/constants/shadowProps";
-import { theme } from "@/src/constants/theme";
-import responsiveToPx, { responsiveToPxByHeight } from "@/src/utils/responsiveToPx";
+import ChatHeader from "../components/ChatHeader";
 import AiChatBox from "../components/AiChatBox";
 import UserChatBox from "../components/UserChatBox";
+import ChatInput from "../components/ChatInput";
 import { useChatting } from "../hooks/useChatting";
 import type { TThreadDate } from "../types/chattingTypes";
-import ChatHeader from "../components/ChatHeader";
+import { responsiveToPxByHeight } from "@/src/utils/responsiveToPx";
 
 type ChattingContainerProps = {
   threadDate: TThreadDate;
@@ -80,24 +78,9 @@ export default function ChattingContainer({
             )
           )}
         </ScrollBox>
-        <KeyboardAvoidingBox behavior={Platform.OS === "ios" ? "padding" : "height"}>
-          {!readonly && (
-            <ChatInputBox>
-              <ChatInput
-                placeholder="오늘 하루에 대해 말해주세요"
-                multiline={true}
-                value={input}
-                onChangeText={(e) => setInput(e)}
-                hitSlop={15}
-                placeholderTextColor={theme.colors.placeholderText}
-              />
-              <ChatButton hitSlop={15} style={shadowProps} onPress={handleSubmitPress}>
-                <ButtonImage source={require("@/assets/images/chatButton.png")} />
-              </ChatButton>
-            </ChatInputBox>
-          )}
-        </KeyboardAvoidingBox>
+        <ChatInput input={input} setInput={setInput} onSubmitPress={handleSubmitPress} readonly={readonly} />
       </SafeView>
+      <InfoText>뒤로 가기 버튼을 누르면 대화가 자동으로 요약됩니다</InfoText>
     </Fragment>
   );
 }
@@ -112,44 +95,10 @@ const ScrollBox = styled(ScrollView)`
   background-color: ${({ theme }) => theme.colors.whiteBlue};
 `;
 
-const KeyboardAvoidingBox = styled.KeyboardAvoidingView`
-  width: 100%;
-  justify-content: center;
-  align-items: center;
-`;
-
-const ChatInputBox = styled.View`
-  width: 100%;
-  min-height: ${responsiveToPx("100px")};
-  background-color: ${({ theme }) => theme.colors.whiteBlue};
-  flex-direction: row;
-  justify-content: center;
-  align-items: flex-end;
-  gap: ${({ theme }) => theme.gap.md};
-  padding-bottom: ${responsiveToPx("40px")};
-`;
-
-const ChatInput = styled.TextInput`
-  width: ${responsiveToPx("333px")};
-  max-height: ${responsiveToPx("100px")};
-  padding: ${responsiveToPx("11px")} ${responsiveToPx("16px")};
-  background-color: ${({ theme }) => theme.colors.white};
-  border: 1px solid ${({ theme }) => theme.colors.gray};
-  border-radius: ${({ theme }) => theme.borderRadius.sm};
+const InfoText = styled.Text`
+  text-align: center;
+  bottom: ${responsiveToPxByHeight("30px")};
+  color: ${({ theme }) => theme.colors.assistantChat};
   font-family: ${({ theme }) => theme.fontFamily.nsRegular};
-  font-size: ${({ theme }) => theme.fontSize.base};
-`;
-
-const ChatButton = styled.TouchableOpacity`
-  width: ${responsiveToPx("44px")};
-  height: ${responsiveToPx("44px")};
-  border-radius: 9999px;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-`;
-
-const ButtonImage = styled(Img)`
-  width: 120%;
-  height: 120%;
+  font-size: ${({ theme }) => theme.fontSize.xs};
 `;
